+++ packages/components/radio/src/use-radio-group.ts
@@ -1,4 +1,8 @@
-import { type ComponentPropsWithoutRef, createContext, useContext } from "react";
+import {
+  createContext,
+  useContext,
+  type ComponentPropsWithoutRef,
+} from "react";
 
 export type RadioGroupContextType = ComponentPropsWithoutRef<"input"> | null;
 
@@ -6,7 +10,9 @@ export const RadioGroupContext = createContext<RadioGroupContextType>(null);
 export const useRadioGroup = () => {
   const context = useContext(RadioGroupContext);
   if (!context) {
-    throw new Error("RadioGroupItem component must be wrapped inside <RadioGroup />");
+    throw new Error(
+      "RadioGroupItem component must be wrapped inside <RadioGroup />",
+    );
   }
   return context;
 };