+++ apps/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # web
 
+## 0.1.54
+
+### Patch Changes
+
+- @jamsr-ui/react@0.0.52
+
 ## 0.1.53
 
 ### Patch Changes